requirements.txt
google-generativeai
pymongo
python-dotenv
PyMuPDF
faiss-cpu
fitz
frontend